Refactored to reduce code duplication between the CLI sub-classes
