+++ src/main/scala/org/clulab/reach/AnnotationsCLI.scala
@@ -2,8 +2,7 @@ package org.clulab.reach
 
 import util.{Failure, Success, Try}
 import scala.collection.JavaConverters._
-import scala.collection.parallel.ForkJoinTaskSupport
-import scala.io.Source
+
 import com.typesafe.config.ConfigFactory
 import com.typesafe.scalalogging.LazyLogging
 import org.apache.commons.io.{FileUtils, FilenameUtils}
@@ -12,20 +11,7 @@ import java.io.File
 import java.util.Date
 import java.nio.charset.Charset
 import java.nio.charset.StandardCharsets.UTF_8
-import ai.lum.common.FileUtils._
 import ai.lum.common.ConfigUtils._
-import org.clulab.odin._
-import org.clulab.reach.`export`.arizona.ArizonaOutputter
-import org.clulab.reach.`export`.cmu.CMUExporter
-import org.clulab.reach.assembly._
-import org.clulab.reach.assembly.export.{AssemblyExporter, AssemblyRow, ExportFilters}
-import org.clulab.reach.export.OutputDegrader
-import org.clulab.reach.export.fries.FriesOutput
-import org.clulab.reach.export.indexcards.IndexCardOutput
-import org.clulab.reach.export.serial.SerialJsonOutput
-import org.clulab.reach.mentions.CorefMention
-import org.clulab.reach.mentions.serialization.json._
-import org.clulab.reach.utils.MentionManager
 import org.clulab.utils.Serializer
 
 /**
@@ -37,109 +23,21 @@ import org.clulab.utils.Serializer
   *   - Remove assembly functionality (including output formats relying on assembly).
   */
 class AnnotationsCLI(
-  val papersDir: File,
-  val outputDir: File,
-  val statsKeeper: ProcessingStats = new ProcessingStats,
-  val encoding: Charset = UTF_8,
-  val restartFile: Option[File] = None
-) extends LazyLogging {
+  override val papersDir: File,
+  override val outputDir: File,
+  override val statsKeeper: ProcessingStats = new ProcessingStats,
+  override val encoding: Charset = UTF_8,
+  override val restartFile: Option[File] = None
+) extends CLI(papersDir, outputDir, statsKeeper, encoding, restartFile) {
 
-  /** Return a (possibly empty) set of filenames for input file (papers) which have
-      already been successfully processed and which can be skipped. */
-  val skipFiles: Set[String] = restartFile match {
-    case None => Set.empty[String]
-    case Some(f) =>
-      // get set of nonempty lines
-      val lines: Set[String] = f.readString(encoding).split("\n").filter(_.nonEmpty).toSet
-      lines
-  }
+
 
   /** Lock file object for restart file. */
   private val restartFileLock = new AnyRef
 
-  /** In the restart log file, record the given file as successfully completed. */
-  def fileSucceeded (file: File): Unit = if (restartFile.nonEmpty) {
-    restartFileLock.synchronized {
-      restartFile.get.writeString(
-        string = s"${file.getName}\n",
-        charset = encoding,
-        append = true,
-        gzipSupport = false
-      )
-    }
-  }
-
-  def reportException(file: File, e: Throwable): Unit = {
-    val filename = file.getName
-    val paperID = FilenameUtils.removeExtension(filename)
-    val report =
-      s"""
-         |==========
-         |
-         | ¡¡¡ NxmlReader error !!!
-         |
-         |paper: $paperID
-         |
-         |error:
-         |${e.toString}
-         |
-         |stack trace:
-         |${e.getStackTrace.mkString("\n")}
-         |
-         |==========
-         |""".stripMargin
-    logger.error(report)
-  }
-
-  /** Process papers **/
-  def preAnnotatePapers(threadLimit: Option[Int]): Int = {
-    logger.info("Initializing Reach ...")
-
-    val files = papersDir.listFilesByRegex(pattern=ReachInputFilePattern, caseInsensitive = true, recursive = true).toVector.par
-
-    // limit parallelization
-    if (threadLimit.nonEmpty) {
-      files.tasksupport =
-        new ForkJoinTaskSupport(new scala.concurrent.forkjoin.ForkJoinPool(threadLimit.get))
-    }
-
-    val errorCounts = for {
-      file <- files
-      filename = file.getName
-      if ! skipFiles.contains(filename)
-    } yield {
-      val error: Int = try {
-        // Count the number of failed files, not failed formats.
-        math.signum(preAnnotatePaper(file))
-      } catch {
-        case e: Exception =>
-          // The reading itself, rather than the format, could have failed.
-          reportException(file, e)
-          1
-      }
-      error
-    }
-    val paperCount = errorCounts.length
-    val errorCount = errorCounts.sum
-    val message =
-      if (errorCount > 0)
-        s"Reach encountered $errorCount error(s) with the $paperCount paper(s).  Please check the log."
-      else
-        s"Reach encountered no errors with the $paperCount paper(s)."
-    logger.info(message)
-    errorCount
-  }
-
-  def prepareMentionsForMITRE (mentions: Seq[Mention]): Seq[CorefMention] = {
-    // NOTE: We're already doing this in the exporter, but the mentions given to the
-    // Assembler probably need to match since flattening results in a loss of information
-    OutputDegrader.prepareForOutput(mentions)
-  }
-
-  def doAssembly (mns: Seq[Mention]): Assembler = Assembler(mns)
 
   // Returns count of outputFormats which errored.
-  def preAnnotatePaper(file: File): Int = {
+  override def processPaper(file: File, withAssembly:Boolean = false): Int = {
     val paperId = FilenameUtils.removeExtension(file.getName)
     val startNS = System.nanoTime
     val startTime = AnnotationsCLI.now
@@ -186,10 +84,6 @@ class AnnotationsCLI(
     errorCount
   }
 
-
-
-  /** Return the duration, in seconds, between the given nanosecond time values. */
-  private def durationToS (startNS:Long, endNS:Long): Long = (endNS - startNS) / 1000000000L
 }
 
 
@@ -267,5 +161,5 @@ object RunAnnotationsCLI extends App with LazyLogging {
     restartFile = if (useRestart) Some(restartFile) else None
   )
 
-  cli.preAnnotatePapers(Some(threadLimit))
+  cli.processPapers(Some(threadLimit), withAssembly = false)
 }
